server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
rembg>=2.0.0
pillow>=10.0.0